First try at implementing Bullet Physics, if you look into the sky you can see three boxes flying down. As a sort of Reminder for us, here is how it works:
ShapeInfos contains all possible collision shapes and CollisionShapeProvider builds them based on the ShapeInfo (e.g. Shape Info would finally be LEVEL01 or WEAPON_SHOTGUN).

Then we have ObjectTypes and ModelViewState, both are responsible for drawing the model on the Scene based on the ObjectType (basically the same as ShapeInfo, but for Meshes).

Lastly there is GameEntities which is responsible for Creating an Entity (which in turns makes the Systems outlined above create collisionshapes and meshes)